AcadSpace

+++ app/Container/Acadspace/src/Controllers/CalendarioController.php
@@ -18,8 +18,6 @@ use Yajra\DataTables\DataTables;
 
 class CalendarioController extends Controller
 {
-
-
     /**
      * Funcion mostrar vista de gestion aulas - calendario
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
@@ -32,9 +30,9 @@ class CalendarioController extends Controller
     /**
      * Funcion para cargar un select en la vista de acuerdo a la informacion
      * recibida de otro select que viene en $espacio
-     * @param \Illuminate\Http\Request $request
-     * @param varchar $espacio
-     * @return \Illuminate\Http\JsonResponse
+     * @param Request $request
+     * @param $espacio
+     * @return \Illuminate\Http\JsonResponse | \App\Container\Overall\Src\Facades\AjaxResponse
      */
     public function cargarSalasCalendario(Request $request, $espacio)
     {
@@ -43,37 +41,43 @@ class CalendarioController extends Controller
                 ->get();
             return response()->json($aula);
         }
+        return AjaxResponse::fail(
+            '¡Lo sentimos!',
+            'No se pudo completar tu solicitud.'
+        );
     }
 
     /**
      * Retorno un arreglo tipo JSON con todos los eventos actuales en
      * el calendario de acuerdo a la sala seleccionada
-     * @param \Illuminate\Http\Request $request
+     * @param Request $request
      * @return array
      */
     public function cargaEventos(Request $request)
     {
-        $sala = $request['sala'];
-        $data = array();
-        $id = Calendario::where('CAL_Sala', '=', $sala)->pluck('PK_CAL_id');
-        $titulo = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Titulo');
-        $color = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Color');
-        $fecha_inicial = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Fecha_Ini');
-        $fecha_final = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Fecha_Fin');
-        $count = count($id);
+        if ($request->ajax() && $request->isMethod('POST')) {
+            $sala = $request['sala'];
+            $data = array();
+            $id = Calendario::where('CAL_Sala', '=', $sala)->pluck('PK_CAL_id');
+            $titulo = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Titulo');
+            $color = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Color');
+            $fecha_inicial = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Fecha_Ini');
+            $fecha_final = Calendario::where('CAL_Sala', '=', $sala)->pluck('CAL_Fecha_Fin');
+            $count = count($id);
 
-        for ($i = 0; $i < $count; $i++) {
-            $data[$i] = array(
-                "title" => $titulo[$i],
-                "start" => $fecha_inicial[$i],
-                "end" => $fecha_final[$i],
-                "backgroundColor" => $color[$i],
-                "id" => $id[$i],
+            for ($i = 0; $i < $count; $i++) {
+                $data[$i] = array(
+                    "title" => $titulo[$i],
+                    "start" => $fecha_inicial[$i],
+                    "end" => $fecha_final[$i],
+                    "backgroundColor" => $color[$i],
+                    "id" => $id[$i],
 
-            );
+                );
+            }
+            json_encode($data);
+            return $data;
         }
-        json_encode($data);
-        return $data;
 
     }
 
@@ -81,27 +85,29 @@ class CalendarioController extends Controller
     /**
      *Funcion para crear evento en el calendario
      * no retorna nada, la validacion de que se creo
-     * se hace en el formulario
+     * se hace en el formulario     *
      */
-    public function create()
+    public function create(Request $request)
     {
-        $titulo = $_POST['title'];
-        $fecha_inicio = $_POST['start'];
-        // $fecha_fin = date($fecha_inicio, (strtotime ("+2 Hours")));
-        $fecha_fin = strtotime('+2 hours', strtotime($fecha_inicio));
-        $fecha_final = date('Y-m-d H:i:s', $fecha_fin);
-        $color = $_POST['background'];
-        $sala = $_POST['salaSeleccionada'];
-        $evento = new Calendario();
+        if ($request->ajax() && $request->isMethod('POST')) {
+            $titulo = $request['title'];
+            $fecha_inicio = $request['start'];
+            // $fecha_fin = date($fecha_inicio, (strtotime ("+2 Hours")));
+            $fecha_fin = strtotime('+2 hours', strtotime($fecha_inicio));
+            $fecha_final = date('Y-m-d H:i:s', $fecha_fin);
+            $color = $request['background'];
+            $sala = $request['salaSeleccionada'];
+            $evento = new Calendario();
 
-        $evento->CAL_Titulo = $titulo;
-        $evento->CAL_Fecha_Ini = $fecha_inicio;
-        $evento->CAL_Fecha_Fin = $fecha_final;
-        $evento->CAL_Color = $color;
-        $evento->CAL_Sala = $sala;
+            $evento->CAL_Titulo = $titulo;
+            $evento->CAL_Fecha_Ini = $fecha_inicio;
+            $evento->CAL_Fecha_Fin = $fecha_final;
+            $evento->CAL_Color = $color;
+            $evento->CAL_Sala = $sala;
 
 
-        $evento->save();
+            $evento->save();
+        }
     }
 
 
@@ -110,46 +116,46 @@ class CalendarioController extends Controller
      * no retorna nada, la validacion de que se actualizo
      * se hace en el formulario
      */
-    public function update()
+    public function update(Request $request)
     {
+        if ($request->ajax() && $request->isMethod('POST')) {
+            //Valores recibidos via ajax
+            $id = $request['id'];
+            $title = $request['title'];
+            $start = $request['start'];
+            $end = $request['end'];
+            $allDay = $request['allday'];
+            $back = $request['background'];
+            $evento = Calendario::find($id);
+            if (empty($end)) {
+            } else {
+                $evento->CAL_Fecha_Fin = $end;
+            }
+            $evento->CAL_Fecha_Ini = $start;
+            $evento->CAL_Color = $back;
+            $evento->CAL_Titulo = $title;
 
-        //Valores recibidos via ajax
-        $id = $_POST['id'];
-        $title = $_POST['title'];
-        $start = $_POST['start'];
-        $end = $_POST['end'];
-        $allDay = $_POST['allday'];
-        $back = $_POST['background'];
-
-        $evento = Calendario::find($id);
-        if (empty($end)) {
-        } else {
-            $evento->CAL_Fecha_Fin = $end;
+            $evento->save();
         }
-        $evento->CAL_Fecha_Ini = $start;
-        $evento->CAL_Color = $back;
-        $evento->CAL_Titulo = $title;
-
-        $evento->save();
     }
 
 
     /**
      *Funcion para eliminar un evento del calendario
      */
-    public function delete()
-    {
-        $id = $_POST['id'];
-
-        Calendario::destroy($id);
+    public function delete(Request $request)    {
+        if ($request->ajax() && $request->isMethod('POST')) {
+            $id = $request['id'];
+            Calendario::destroy($id);
+        }
     }
 
     /**
      * Funcion que carga el datatable con solicitudes aprobadas de acuerdo
      * a la sala seleccionada
-     * @param \Illuminate\Http\Request $request
-     * @param id $sala
-     * @return \Illuminate\Http\Response | \Yajra\DataTables\DataTables
+     * @param Request $request
+     * @param $sala
+     * @return \App\Container\Overall\Src\Facades\AjaxResponse | \Yajra\DataTables\
      */
     public function data(Request $request, $sala)
     {
